+++ RecipeStorage/Features/Recipes/Recipe/Details/02_MacrosView/Details/MacroDetailView.swift
@@ -10,7 +10,7 @@ import SwiftUI
 struct MacroDetailView: View {
     
     let macroName: String
-    let macroAmount: Int
+    let macroAmount: Double
     let color: Color
     
     var body: some View {
